requirements.txt
fastapi
uvicorn
requests
python-multipart
openai
python-dotenv
google-generativeai
groq
pymongo